Change: Move interfaces to ./interfaces.ts. Introduce FormattedPendingAmounts interface.
